+++ src/etsy_listing_a_/index.d.ts
@@ -1,3 +1,0 @@
-import type { etsy_listing_T } from '../_types'
-export declare function etsy_listing_a_():Promise<etsy_listing_T[]>
-export { etsy_listing_a_ as _etsy_listing_a1, }