+++ src/loading__s3__etsy_listing_image_a/index.d.ts
@@ -1,0 +1,14 @@
+import type { val__be_atom_triple_T } from '@ctx-core/nanostores'
+import type { Ctx } from '@ctx-core/object'
+import type { etsy_listing_image_T } from '../_types'
+export declare const [
+	loading__s3__etsy_listing_image_a$_,
+	loading__s3__etsy_listing_image_a_,
+	loading__s3__etsy_listing_image_a__set,
+]:val__be_atom_triple_T<etsy_listing_image_T[]|undefined>
+export {
+	loading__s3__etsy_listing_image_a$_ as loading__s3__etsy_listing_image_a__,
+	loading__s3__etsy_listing_image_a$_ as loading_s3_etsy_listing_image_a__,
+	loading__s3__etsy_listing_image_a$_ as loading_s3_etsy_listing_image_a$_,
+}
+export default function loading__s3__etsy_listing_image_a__reload(ctx:Ctx):Promise<void>