+++ src/s3__etsy_listing_a__/index.d.ts
@@ -1,8 +1,0 @@
-import type { ReadableAtom_ } from '@ctx-core/nanostores'
-import type { Be } from '@ctx-core/object'
-import type { etsy_listing_T } from '../_types'
-export declare const s3__etsy_listing_a__:Be<ReadableAtom_<etsy_listing_T[]|undefined>>
-export {
-	s3__etsy_listing_a__ as s3_etsy_listing_a__,
-	s3__etsy_listing_a__ as s3_etsy_listing_a$_,
-}