+++ src/s3__etsy_listing_image_a_/index.d.ts
@@ -1,6 +1,0 @@
-import type { etsy_listing_image_T } from '../_types'
-import type { s3__etsy_listing_image_a__fetch_get__params_T } from '../s3__etsy_listing_image_a__fetch_get'
-export declare function s3__etsy_listing_image_a_(
-	params?:s3__etsy_listing_image_a__fetch_get__params_T
-):Promise<etsy_listing_image_T[]>
-export { s3__etsy_listing_image_a_ as _s3_etsy_listing_image_a, }